added other changes

+++ src/main/java/com/sagar/firstJobApp/job/Impl/JobServiceImpl.java
@@ -27,7 +27,7 @@ public class JobServiceImpl implements JobService {
 
     @Override
     public void createJob(Job job) {
-        job.setId(nextId++);
+//        job.setId(nextId++);
 //        jobs.add(job);
         jobRepository.save(job);
 
@@ -40,12 +40,11 @@ public class JobServiceImpl implements JobService {
 
     @Override
     public boolean deleteJobById(Long id) {
-        try {
+        if (jobRepository.existsById(id)){
             jobRepository.deleteById(id);
             return true;
-        } catch (Exception e) {
-            return false;
         }
+        return false;
     }
 
     @Override
